[ADD] WaterLevelSensor값 저장 -8

+++ src/main/java/capjjangdol/mallangkongth/domain/rearing/WaterNote.java
@@ -20,5 +20,5 @@ public class WaterNote {
     private Pet pet;
 
     @Column(name = "WATER_AMOUNT")
-    private Long amount;
+    private Integer amount;
 }
